STABLE
This is the first stand up of our system fully working E2E.  This is a
usable place to do the demo.  But i am going to try and get checkout
hooked up as well, as i think that would really show the core MVP
working... hows that for a commit message.